backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==1.3.0
marshmallow==3.20.1
python-dateutil==2.8.2 
gunicorn==21.2.0
werkzeug==3.1.3
flask-login==0.6.3
PyJWT==2.8.0
flask-mail==0.9.1
